requirements.txt
aiofiles==23.2.1
aiogram==3.13.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==4.2.16
django-bootstrap5==22.2
django-templated-mail==1.1.1
django_debug_toolbar==3.8.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
Faker==12.0.1
flake8==5.0.4
flake8-docstrings==1.7.0
frozenlist==1.4.1
idna==3.8
iniconfig==2.0.0
isort==5.13.2
magic-filter==1.0.12
mccabe==0.7.0
mixer==7.2.2
multidict==6.1.0
oauthlib==3.2.2
packaging==23.0
pep8-naming==0.13.3
Pillow==9.3.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydocstyle==6.3.0
pyflakes==2.5.0
PyJWT==2.9.0
pytest==7.1.3
pytest-django==4.5.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
snowballstemmer==2.2.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yapf==0.32.0
yarl==1.11.1
